% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{compute_correlation}
\alias{compute_correlation}
\title{Compute sample correlation}
\usage{
compute_correlation(
  study,
  transform = c("vst", "log2-tpm"),
  method = c("pearson", "spearman")
)
}
\arguments{
\item{study}{A \code{rsemflow_study} or study path.}

\item{transform}{\code{vst} or \code{log2-tpm}.}

\item{method}{\code{pearson} or \code{spearman}.}
}
\value{
A sample-by-sample correlation matrix.
}
\description{
Compute sample correlation
}
