\name{cli_main}
\alias{cli_main}
\title{RSEMflow command-line entry point}
\usage{
cli_main(args = commandArgs(trailingOnly = TRUE))
}
\arguments{\item{args}{Command-line argument vector.}}
\value{Returns zero invisibly on success.}
\description{Dispatches the \code{rsemflow <module> <action>} command-line interface.}
